+++ lib/lite_state.dart
@@ -87,8 +87,7 @@ void _lazilyInitializeController(String typeKey) {
   final initializer = _lazyControllerInitializers[typeKey];
   _controllers[typeKey] = initializer!();
   _controllers[typeKey]!.rebuild();
-  debugPrint(
-      'LiteState: LAZILY INITIALIZED CONTROLLER: ${_controllers[typeKey]}');
+  debugPrint('LiteState: LAZILY INITIALIZED CONTROLLER: ${_controllers[typeKey]}');
 }
 
 void _addTemporaryController<T>(
@@ -118,8 +117,7 @@ T findController<T extends LiteStateController>() {
 
 bool _hasControllerInitializer<T extends LiteStateController>() {
   final typeKey = T.toString();
-  return _controllers.containsKey(typeKey) ||
-      _lazyControllerInitializers.containsKey(typeKey);
+  return _controllers.containsKey(typeKey) || _lazyControllerInitializers.containsKey(typeKey);
 }
 
 typedef LiteStateBuilder<T extends LiteStateController> = Widget Function(
@@ -151,8 +149,7 @@ class LiteState<T extends LiteStateController> extends StatefulWidget {
   State<LiteState> createState() => _LiteStateState<T>();
 }
 
-class _LiteStateState<T extends LiteStateController>
-    extends State<LiteState<T>> {
+class _LiteStateState<T extends LiteStateController> extends State<LiteState<T>> {
   Widget? _child;
   bool _isReady = false;
 
@@ -298,6 +295,10 @@ abstract class LiteStateController<T> {
   static final Map<String, dynamic> _streamControllers = {};
 
   LiteRepo? _liteRepo;
+  
+  LiteRepo? get repo {
+    return _liteRepo;
+  }
 
   /// [useLocalStorage] whether to use a local storage (based on `Hive` or not)
   final bool useLocalStorage;
@@ -357,15 +358,17 @@ abstract class LiteStateController<T> {
     String key,
     List<TGenericType> values,
   ) async {
-    await setPersistentValue(key, values);
+    if (!useLocalStorage) {
+      return;
+    }
+    _liteRepo?.setList<TGenericType>(key, values);
   }
 
   List<TGenericType>? getPersistentList<TGenericType>(String key) {
-    final value = getPersistentValue(key);
-    if (value is List) {
-      return value.cast<TGenericType>().toList();
+    if (!useLocalStorage) {
+      return null;
     }
-    return null;
+    return _liteRepo?.getList<TGenericType>(key);
   }
 
   Future setPersistentValue<TType>(